refactor main to another file and remove garbage
