+++ Week 2/Day 2/Submissions/aleksandr_safin/run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python3 run.py
